set environment variable, use master key, store aws creds

+++ ansible-master/scripts/does-master.py
@@ -12,7 +12,9 @@ from state import State
 OUT_FOLDER = "~/.does-master"
 LOG_FOLDER = f"{OUT_FOLDER}/logs"
 STATE_PATH = f"{OUT_FOLDER}/state.json"
-RESULTS_ZIP_PATH = "/tmp/results.zip"
+RESULTS_ZIP_PATH = "/tmp/results"
+DOES_PRJ_DIR_VARIABLE = "DOES_PROJECT_DIR"
+
 
 def parse_arguments():
     parser = argparse.ArgumentParser(description="Utility to run benchmarks on AWS using the DoE-Suite")
@@ -70,7 +72,7 @@ def handle_ansible_cmd(args, state):
     commit = args.commit
 
     # TODO: change project ID to commit hash -> add functionality to kill old benchmarks
-    # subprocess.run(["poetry", "run", "ansible-playbook", "src/experiment-suite.yml", "-e", f"suite={benchmark} id=new"], cwd=self.state["home"])
+    subprocess.run(["poetry", "run", "ansible-playbook", "src/experiment-suite.yml", "-e", f"suite={benchmark} id=new"], cwd=state.state["home"])
 
     # TODO: get result path fron ETL pipeline!
     path = ""
@@ -89,7 +91,8 @@ def handle_results_cmd(args, state):
     state.update()
 
     if do_list_results:
-        for result in self.state["results"]:
+        print("The following results are available:")
+        for result in state.state["results"]:
             print(f"\t-", result)
         return
 
@@ -97,7 +100,7 @@ def handle_results_cmd(args, state):
 
     results_path = tempfile.mkdtemp()
 
-    for result in self.state["results"]:
+    for result in state.state["results"]:
         if result["path"] in files_to_fetch:
             commit_folder = f"{results_path}/{result['commit']}"
             create_folder(commit_folder)
@@ -105,16 +108,22 @@ def handle_results_cmd(args, state):
             # TODO: find better way to not overwrite result files
             shutil.copyfile(result["path"], f"{commit_folder}/{results_path.split('/')[1]}")
 
-    if os.path.exists(RESULTS_ZIP_PATH):
-        os.remove(RESULTS_ZIP_PATH)
+    if len(state.state["results"]) > 0:
+        if os.path.exists(RESULTS_ZIP_PATH):
+            os.remove(RESULTS_ZIP_PATH)
 
-    shutil.make_archive(RESULTS_ZIP_PATH, "zip", results_path)
+        shutil.make_archive(RESULTS_ZIP_PATH, "zip", results_path)
 
-    print(f"Download the requested files from {RESULTS_ZIP_PATH} (e.g., using scp).")
+        print(f"Download the requested files from {RESULTS_ZIP_PATH} (e.g., using scp).")
 
 if __name__ == '__main__':
     setup_outdir()
-    state = State(STATE_PATH, os.environ["DOES_PROJECT_DIR"])
+
+    if DOES_PRJ_DIR_VARIABLE not in os.environ:
+        logging.error(f"The variable '{DOES_PRJ_DIR_VARIABLE}' is not set. Set it to the directory of the doe-suite.")
+        exit(1)
+
+    state = State(STATE_PATH, f"{os.environ[DOES_PRJ_DIR_VARIABLE]}/doe-suite")
 
     args = parse_arguments()
